admin/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
python-dotenv==1.0.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.128.0
gspread==6.1.2
gunicorn==22.0.0
werkzeug==3.0.3
